requirements.txt
langgraph
langchain-google-genai
ccxt
pandas
ta
feedparser
beautifulsoup4
python-telegram-bot
apscheduler
python-dotenv
